refactor: jest to vitest migration
